% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_w_next.R
\name{diff_w_next}
\alias{diff_w_next}
\title{Вычисляет разности между текущими и предыдущими элементами в векторе с сохранением первого элемента}
\usage{
diff_w_next(x)
}
\arguments{
\item{x}{Вектор чисел или вектор, для которого необходимо вычислить разницы.}
}
\value{
Вектор той же длины, что и \code{x}. Первый элемент совпадает с исходным,
остальные — разницы между элементами массива.
}
\description{
Функция \code{diff_w_next} возвращает вектор, где первый элемент такой же, как и в исходном,
а остальные — разницы между текущим элементом и предыдущим. Используется в пакете skigraph2 для того,
чтобы находить время преодоления спортсменами отдельных отрезков
}
\examples{
x <- c(10, 12, 15, 20)
diff_w_next(x)
# Результат: 10, 2, 3, 5


}
