% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_2025_descr.R
\docType{data}
\name{results_2025}
\alias{results_2025}
\title{Таблица содержит результаты соревнований в лыжных гонках за сезон 2025 в категориях ЭКР, ФКР, ЧР}
\format{
Таблица с 13691 строкой и 28 переменными:
\describe{
\item{\code{place}}{Место участника в итоговом зачёте}
\item{\code{bib}}{Стартовый номер}
\item{\code{athlete}}{ФИО спортсмена}
\item{\code{reg}}{Регион, представляемый спортсменом}
\item{\code{res}}{Результат}
\item{\code{diff}}{Разница во времени с лидером, например, "+0:00.90"}
\item{\code{split}}{Дистания промежуточной отсечки}
\item{\code{comp_id}}{Идентификатор соревнования}
\item{\code{Дата}}{Дата проведения соревнования}
\item{\code{Сп. Дисциплина}}{Спортивная дисциплина}
\item{\code{Категория}}{Категория соревнования}
\item{\code{Пол}}{Пол спортсмена}
\item{\code{event_id}}{Идентификатор события}
\item{\code{Событие}}{Название события или этапа, к которому относится соревнование}
\item{\code{Город}}{Город проведения соревнования}
\item{\code{Регион}}{Регион проведения соревнования}
\item{\code{split_km}}{Длина промежуточного сегмента в километрах}
\item{\code{split_m}}{Длина промежуточного сегмента в метрах}
}
}
\usage{
results_2025
}
\description{
Набор данных содержит результаты российских спортсменов в лыжных гонках данные загружены с сайта flgr-results.ru
}
\details{
Результаты соревнований в российских лыжных гонках за сезон 2025, содержащие промежуточные результаты на отсечках по данным live-трансляций сайта flgr-results.ru за исключением соревнований в лыжероллерах, кроссах, финалов спринта, эстафет, командных соревнований в соревнованиях категорий Этап Кубка России(ЭКР), Финал Кубка России (ФКР), Чемпионат России (ЧР)
}
\examples{
str(results_2025)
}
\keyword{datasets}
