% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R, R/results_rus_descr.R
\docType{data}
\name{results_rus}
\alias{results_rus}
\title{Таблица содержит результаты соревнований в лыжных гонках}
\format{
Таблица с 393247 строк и 28 переменными:
\describe{
\item{\code{Место}}{итоговое место (например, "1", "2", "3") }
\item{\code{Номер}}{стартовый номер}
\item{\code{RUS код}}{RUS код спортсмена}
\item{\code{Спортсмен}}{Имя и фамилия спортсмена}
\item{\code{Г.Р.}}{Год рождения}
\item{\code{Звание}}{Спортивный разряд участника}
\item{\code{Субьект РФ}}{Регион или субъект РФ}
\item{\code{Результат}}{Итоговый результат в минутах, при пропущенных значениях NA}
\item{\code{Отставание}}{Отставание от лидера, может быть NA}
\item{\code{Рус пункты}}{Рус пункты по итогу соревнования}
\item{\code{Очки}}{Очки по итогу соревнования}
\item{\code{link}}{URL-ссылка на спортивное событие, необходимо добавить вначале "https://flgr-results.ru"}
\item{\code{Дата}}{Дата проведения в формате "DD.MM.YYYY"}
\item{\code{Код}}{Идентификатор события или соревнования}
\item{\code{Сп. Дисциплина}}{Название дисциплины}
\item{\code{Категория}}{Категория соревнования}
\item{\code{Пол}}{Пол ("М" или "Ж")}
\item{\code{Примечание}}{Примечания об изменениях в регламенте}
\item{\code{link.x}}{Ссылка на результаты конкретного соревнования, добавить "flgr-results.ru"}
\item{\code{Категория_полн}}{Полное название категории соревнования}
\item{\code{Город}}{Город проведения}
\item{\code{Регион}}{Регион или область проведения}
\item{\code{year}}{Год из даты проведения соревнования}
\item{\code{month}}{Месяц из даты проведения соревнования}
\item{\code{season}}{Сезон (если месяц <= 6, сезон = year, иначе сезон =  year+1)}
\item{\code{dist}}{Длина дистанции из названия соревнования (5, 10, 15, ..., или Спринт)}
\item{\code{start_type}}{Индивидуальный или масс-старт}
\item{\code{style}}{Стиль передвижения, КЛ - классика или СВ - свободный (коньковый)}
\item{\code{type}}{тип - дистанция или спринт}
}

Таблица с 393247 строк и 28 переменными:
\describe{
\item{\code{Место}}{итоговое место (например, "1", "2", "3") }
\item{\code{Номер}}{стартовый номер}
\item{\code{RUS код}}{RUS код спортсмена}
\item{\code{Спортсмен}}{Имя и фамилия спортсмена}
\item{\code{Г.Р.}}{Год рождения}
\item{\code{Звание}}{Спортивный разряд участника}
\item{\code{Субьект РФ}}{Регион или субъект РФ}
\item{\code{Результат}}{Итоговый результат в минутах, при пропущенных значениях NA}
\item{\code{Отставание}}{Отставание от лидера, может быть NA}
\item{\code{Рус пункты}}{Рус пункты по итогу соревнования}
\item{\code{Очки}}{Очки по итогу соревнования}
\item{\code{link}}{URL-ссылка на спортивное событие, необходимо добавить вначале "https://flgr-results.ru"}
\item{\code{Дата}}{Дата проведения в формате "DD.MM.YYYY"}
\item{\code{Код}}{Идентификатор события или соревнования}
\item{\code{Сп. Дисциплина}}{Название дисциплины}
\item{\code{Категория}}{Категория соревнования}
\item{\code{Пол}}{Пол ("М" или "Ж")}
\item{\code{Примечание}}{Примечания об изменениях в регламенте}
\item{\code{link.x}}{Ссылка на результаты конкретного соревнования, добавить "flgr-results.ru"}
\item{\code{Категория_полн}}{Полное название категории соревнования}
\item{\code{Город}}{Город проведения}
\item{\code{Регион}}{Регион или область проведения}
\item{\code{year}}{Год из даты проведения соревнования}
\item{\code{month}}{Месяц из даты проведения соревнования}
\item{\code{season}}{Сезон (если месяц <= 6, сезон = year, иначе сезон =  year+1)}
\item{\code{dist}}{Длина дистанции из названия соревнования (5, 10, 15, ..., или Спринт)}
\item{\code{start_type}}{Индивидуальный или масс-старт}
\item{\code{style}}{Стиль передвижения, КЛ - классика или СВ - свободный (коньковый)}
\item{\code{type}}{тип - дистанция или спринт}
}
}
\usage{
results_rus

results_rus
}
\description{
Набор данных результатов российских спортсменов в лыжных гонках, за исключением летних видов соренований (кросс, лыжероллеры) и финалов спринта, загруженных с сайта flgr-results.ru, начиная с 2009 года

Набор данных результатов российских спортсменов в лыжных гонках, за исключением летних видов соренований (кросс, лыжероллеры) и финалов спринта, загруженных с сайта flgr-results.ru, начиная с 2009 года
}
\details{
Данные собраны с сайта flgr-results.ru, с 2009 по 2025 год
Набор данных результатов российских соревнований в лыжных гонках, за исключением летних видов соренований (кросс, лыжероллеры) и финалов спринта."Не Закончили Дистанцию", "Круговой", "Финалы", "Четверьфинал" и т.д. разбивают файл на части = после "Не Закончили Дистанцию" следует перечисление спортсменов и т.п.

Данные собраны с сайта flgr-results.ru, с 2009 по 2025 год
Набор данных результатов российских соревнований в лыжных гонках, за исключением летних видов соренований (кросс, лыжероллеры) и финалов спринта."Не Закончили Дистанцию", "Круговой", "Финалы", "Четверьфинал" и т.д. разбивают файл на части = после "Не Закончили Дистанцию" следует перечисление спортсменов и т.п.
}
\examples{
str(results_rus)
str(results_rus)
}
\keyword{datasets}
