% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprints_fis_descr.R
\docType{data}
\name{sprints_fis}
\alias{sprints_fis}
\title{Результаты соревнований в дисциплине спринт FIS}
\format{
Таблица с 13691 строкой и 28 переменными:
\describe{
\item{position}{Место по результатам соревнования}
\item{number}{Стартовый номер}
\item{name}{Фамилия и имя участника}
\item{birth_year}{Год рождения}
\item{country}{Страна участника}
\item{time}{Результат соревнования}
\item{link}{Ссылка на общие результаты гонки}
\item{date}{Дата проведения гонки}
\item{comp}{Название соревнования}
\item{cat}{Категория соревнования}
\item{gend}{Пол участника}
\item{note}{Дополнительные заметки}
\item{event_link}{Ссылка на страницы с деталями события (этапа)}
\item{place}{Место проведения}
\item{nat}{Страна}
\item{event_place}{Место проведения соревнования}
\item{event_title}{Название события}
\item{event_id}{Идентификатор события}
\item{season}{Сезон}
\item{len}{Длина дистанции}
\item{lap_n}{Количество кругов}
\item{climb}{Набор высоты}
\item{n_pdf}{Название PDF-файла результатов}
\item{link_pdf}{Ссылка на PDF-файл результатов}
}
}
\usage{
sprints_fis
}
\description{
Набор данных, содержащий результаты спринтерских гонок по лыжным гонкам, включающий сведения о спортсменах, дистанциях, датах, местах проведения и ссылки на результаты
}
\details{
Данные собраны с сайта fis-ski.com, с 2009 по 2025 год
Набор данных, содержащий данные протоколов в дисциплине спринт, в таблицу добавлены длины дистанций для определения сренесоревновательной скорости. Скорость определялась только в квалификации спринта, так как в финалах не всегда доступны данные о преодолении дистанции, кроме того имеет место тактическая борьба, которая искажает результаты
}
\examples{
str(results_fis)
}
\keyword{datasets}
