% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{results_rus}
\alias{results_rus}
\title{Набор данных результатов российских спортсменов}
\format{
A tibble/data.frame with 404163 строк и 22 переменными:
\describe{
\item{\code{Место}}{character — Место финалиста (например, "1", "2", "3") }
\item{\code{Номер}}{character — Номер участника или стартовый номер}
\item{\code{RUS код}}{character — Уникальный код спортсмена}
\item{\code{Спортсмен}}{character — Имя и фамилия спортсмена}
\item{\code{Г.Р.}}{character — Год рождения}
\item{\code{Звание}}{character — Звание спортсмена (например, "змс", "мсмк") }
\item{\code{Субьект РФ}}{character — Регион или субъект РФ}
\item{\code{Результат}}{character — Итоговый результат, может быть NA}
\item{\code{Отставание}}{character — Отставание от лидера, может быть NA}
\item{\code{Рус пункты}}{character — Очки, набранные спортсменом}
\item{\code{Очки}}{character — Общие очки (часто NA)}
\item{\code{link}}{character — URL-адрес страницы с результатами}
\item{\code{Дата}}{character — Дата проведения в формате "DD.MM.YYYY"}
\item{\code{Код}}{integer — Идентификатор события или соревнования}
\item{\code{Сп. Дисциплина}}{character — Название дисциплины}
\item{\code{Категория}}{character — Категория участника}
\item{\code{Пол}}{character — Пол ("М" или "Ж")}
\item{\code{Примечание}}{character — Дополнительные примечания}
\item{\code{link.x}}{character — Вторая ссылка на результаты}
\item{\code{Категория_полн}}{character — Полная категория соревнования}
\item{\code{Город}}{character — Город проведения}
\item{\code{Регион}}{character — Регион или область проведения}
}
}
\usage{
results_rus
}
\description{
Таблица с информацией о результатах участников российских спортивных соревнований.
}
\details{
Данные собраны из результатов спортивных соревнований, проводимых в России.
Включают информацию о спортсменах, их результатах, регионах и деталях соревнований.
}
\examples{
str(results_rus)
}
\keyword{datasets}
